Move out the helper
