+++ test/enivronmentalist_test.exs
@@ -2,7 +2,6 @@ defmodule EnivronmentalistTest do
   use ExUnit.Case
   doctest Enivronmentalist
 
-  test "the truth" do
-    assert 1 + 1 == 2
+  test "it adds missing entries" do
   end
 end
